pyproject.toml
[tool.poetry]
name = "edict-heart"
version = "0.1.0"
description = "Heart health analysis project in EDICT Lab, IISER Bhopal."
authors = ["Sattwik Sahu <sahusattwik@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
pandas = "^2.2.0"
pyserial = "^3.5"
pyyaml = "^6.0.1"
structlog = "^24.1.0"
typer = "^0.9.0"
pydantic = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
